+++ src/main/java/com/base/basesetup/dto/CostDebitNoteGstDTO.java
@@ -1,5 +1,7 @@
 package com.base.basesetup.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,9 +15,9 @@ public class CostDebitNoteGstDTO {
 	private Long id;
 	private String chargeAcc;
 	private String subLodgerCode;
-	private double dBillAmt;
-	private double crBillAmt;
-	private double dBLCAmt;
-	private double crLCAmt;
+	private BigDecimal dBillAmt;
+	private BigDecimal crBillAmt;
+	private BigDecimal dBLCAmt;
+	private BigDecimal crLCAmt;
 	private String remarks;
 }
