+++ src/main/java/com/base/basesetup/entity/CostInvSummaryVO.java
@@ -1,5 +1,7 @@
 package com.base.basesetup.entity;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -30,21 +32,21 @@ public class CostInvSummaryVO {
 	@Column(name = "costinvsummaryid")
 	private Long id;
 	@Column(name ="totchargesbillcurramt",precision =10,scale =2)
-	private Double totChargesBillCurrAmt;
+	private BigDecimal totChargesBillCurrAmt;
 	@Column(name ="totchargeslcamt",precision =10,scale =2)
-	private Double totChargesLcAmt;
+	private BigDecimal totChargesLcAmt;
 	@Column(name ="actbillcurramt",precision =10,scale =2)
-	private Double actBillCurrAmt;
+	private BigDecimal actBillCurrAmt;
 	@Column(name ="actbilllcamt",precision =10,scale =2)
-	private Double actBillLcAmt;
+	private BigDecimal actBillLcAmt;
 	@Column(name ="netbillcurramt",precision =10,scale =2)
-	private Double netBillCurrAmt;
+	private BigDecimal netBillCurrAmt;
 	@Column(name ="netbilllcamt",precision =10,scale =2)
-	private Double netBillLcAmt;
+	private BigDecimal netBillLcAmt;
 	@Column(name ="roundoff",precision =10,scale =2)
-	private Double roundOff;
+	private BigDecimal roundOff;
 	@Column(name ="gstinputlcamt",precision =10,scale =2)
-	private Double gstInputLcAmt;
+	private BigDecimal gstInputLcAmt;
 	
 	@ManyToOne
 	@JsonBackReference
